Alter algorithm of binary search to linear search.
Such action was necessary due to binary search instability to some instances of the problem. Not only for instance size, but also for wrapped variables such as two types of fetch keys and within a range, not being an exact key.

The algorithm is acceptable depending on the instance of the problem, having complexity O (n) for the worst and medium case, and O (1) for the best, but rare!